requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
asgiref==3.7.2; python_version >= '3.7'
certifi==2023.11.17; python_version >= '3.6'
cffi==1.16.0; python_version >= '3.8'
charset-normalizer==3.3.2; python_version >= '3.7'
cryptography==41.0.7
defusedxml==0.7.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
dj-database-url==2.1.0
dj-email-url==1.0.6
django-allauth==0.58.2
django-cache-url==3.4.5
django-debug-toolbar==4.2.0
django-extensions==3.2.3
django==4.2.8
environs[django]==9.5.0
gunicorn==21.2.0
idna==3.6; python_version >= '3.5'
marshmallow==3.20.1; python_version >= '3.8'
oauthlib==3.2.2; python_version >= '3.6'
packaging==23.2; python_version >= '3.7'
psycopg2-binary==2.9.9
pycparser==2.21
pyjwt[crypto]==2.8.0; python_version >= '3.7'
python-dotenv==1.0.0; python_version >= '3.8'
python3-openid==3.2.0
requests-oauthlib==1.3.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
requests==2.31.0; python_version >= '3.7'
sqlparse==0.4.4; python_version >= '3.5'
typing-extensions==4.9.0; python_version < '3.11'
urllib3==2.1.0; python_version >= '3.8'
whitenoise==6.6.0
